backend/requirements.txt
fastapi
uvicorn
gitpython
sentence-transformers
chromadb
openai
pydantic
python-dotenv
supabase
networkx
graphviz
